Rectored code and removed duplicated brand: brand
