+++ ComponentLibrary/ComponentLibraryHome.swift
@@ -7,7 +7,7 @@ struct ComponentLibraryHome: View {
         ("Links", AnyView(LinkPage()), "rectangle.grid.1x2"),
         ("Tags", AnyView(TagDemoPage()), "rectangle.grid.1x2"),
         ("Separator", AnyView(SeparatorDemoPage()), "rectangle.grid.1x2"),
-        ("Radio & Checkbox", AnyView(InputPage()), "rectangle.grid.1x2"),
+        ("Input", AnyView(InputPage()), "rectangle.grid.1x2"),
         ("Confirmation Page", AnyView(ConfirmationPage()), "dock.rectangle"),
         ("ColorTokenSystem", AnyView(ColorSwatches()), "rectangle.fill.on.rectangle.fill"),
         ("TypographyTokenSystem", AnyView(FontsPage()), "rectangle.fill.on.rectangle.fill"),
