+++ ComponentLibrary/Models/TypographyTokenModels.swift
@@ -7,6 +7,7 @@
 
 import SwiftUI
 
+
 struct TextStyleToken: Decodable {
     let fontName: String
     let weight: String
@@ -29,3 +30,33 @@ struct TypographyTokens: Decodable {
     }
 }
 
+enum MyTextStyle: String {
+    case h1, h2, h3, h4, h5, h6
+    case p1, p2, p3
+    case button
+    case link
+}
+
+
+struct TypographyStyleModifier: ViewModifier {
+    let style: MyTextStyle
+    @Environment(\.brand) private var brand
+
+    func body(content: Content) -> some View {
+        let manager = TypographyTokenManager.shared
+        let font = manager.font(for: brand, styleName: style.rawValue)
+        let kerning = manager.letterSpacing(for: brand, styleName: style.rawValue)
+        let lineSpacing = manager.lineSpacing(for: brand, styleName: style.rawValue)
+
+        return content
+            .font(font)
+            .kerning(kerning)
+            .lineSpacing(lineSpacing)
+    }
+}
+
+extension View {
+    func typographyStyle(_ style: MyTextStyle) -> some View {
+        modifier(TypographyStyleModifier(style: style))
+    }
+}
