+++ ComponentLibrary/Pages/ButtonPage.swift
@@ -12,6 +12,7 @@ struct ButtonView: View {
     @Environment(\.colorScheme) var colorScheme
     @Environment(\.brand) private var brand
     
+    
     var body: some View {
         GeometryReader { geometry in
             ZStack {
