+++ ComponentLibrary/Pages/SeparatorPage.swift
@@ -9,14 +9,14 @@ struct SeparatorDemoPage: View {
        ScrollView {
            VStack(alignment: .leading, spacing: 16) {
                Text("Separator Demo Page")
-                   .typographyStyle(.h1,  brand: brand)
+                   .typographyStyle(.h1)
                    .padding(.bottom, 8)
                
                Text("This is an example of a **horizontal separator**:")
-                   .typographyStyle(.h4,  brand: brand)
+                   .typographyStyle(.h4)
                
                Text("Lorem ipsum dolor sit amet, consectetur adipiscing elit.")
-                   .typographyStyle(.p1,  brand: brand)
+                   .typographyStyle(.p1)
                
                    ButtonComponent(
                 
@@ -27,7 +27,7 @@ struct SeparatorDemoPage: View {
                SeparatorComponent( type: .horizontal, label: "OR")
                
                
-               Text("Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.").typographyStyle(.p1,  brand: brand)
+               Text("Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.").typographyStyle(.p1)
               
                    ButtonComponent(
                        title: "Button",
@@ -41,25 +41,25 @@ struct SeparatorDemoPage: View {
                Spacer().frame(height: 10)
                
                Text("This is an example of a **vertical separator**:")
-                   .typographyStyle(.h4,  brand: brand)
+                   .typographyStyle(.h4)
                
                HStack {
                    // **First Column - Features**
                    VStack(spacing: 8) {
                        Text("Features")
-                           .typographyStyle(.h5, brand: brand)
+                           .typographyStyle(.h5)
                            .frame(maxWidth: .infinity, alignment: .center)
                        
                        Text("✓ High Performance")
-                           .typographyStyle(.p1, brand: brand)
+                           .typographyStyle(.p1)
                            .frame(maxWidth: .infinity, alignment: .center)
                        
                        Text("✓ Easy Integration")
-                           .typographyStyle(.p1, brand: brand)
+                           .typographyStyle(.p1)
                            .frame(maxWidth: .infinity, alignment: .center)
                        
                        Text("✓ Secure & Reliable")
-                           .typographyStyle(.p1, brand: brand)
+                           .typographyStyle(.p1)
                            .frame(maxWidth: .infinity, alignment: .center)
                    }
                    
@@ -67,31 +67,31 @@ struct SeparatorDemoPage: View {
                    
                    VStack(spacing: 8) {
                        Text("Pricing")
-                           .typographyStyle(.h5, brand: brand)
+                           .typographyStyle(.h5)
                            .frame(maxWidth: .infinity, alignment: .center)
                        
                        Text("$9.99 / month")
-                           .typographyStyle(.p1, brand: brand)
+                           .typographyStyle(.p1)
                            .frame(maxWidth: .infinity, alignment: .center)
                        
                        Text("$99.99 / year")
-                           .typographyStyle(.p1, brand: brand)
+                           .typographyStyle(.p1)
                            .frame(maxWidth: .infinity, alignment: .center)
                        
                        Text("Custom Enterprise Plans")
-                           .typographyStyle(.p1, brand: brand)
+                           .typographyStyle(.p1)
                            .frame(maxWidth: .infinity, alignment: .center)
                    }
                }
                Spacer().frame(height: 10)
                Text("Vertical separator with Label:")
-                   .typographyStyle(.h4,  brand: brand)
+                   .typographyStyle(.h4)
                
                HStack {
                    
                    VStack(spacing: 8) {
                        Text("Sign In with Email")
-                           .typographyStyle(.h5, brand: brand)
+                           .typographyStyle(.h5)
                            .frame(maxWidth: .infinity, alignment: .center)
                        
                        ButtonComponent(
@@ -105,7 +105,7 @@ struct SeparatorDemoPage: View {
                    
                    VStack(spacing: 8) {
                        Text("Sign In with Social Media")
-                           .typographyStyle(.h5, brand: brand)
+                           .typographyStyle(.h5)
                            .frame(maxWidth: .infinity, alignment: .center)
                        
                        ButtonComponent(
